student-course-center/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "student-course-center"
version = "0.1.0"
description = "Console client for StudentCourseCenterDb (SQL Server)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyodbc>=5.0.0",
    "python-dotenv>=1.0.0",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "Faker>=24.0.0"]

[project.scripts]
scc = "student_course_center.console.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
